+++ src/constants.ts
@@ -48,6 +48,7 @@ export const HFP_PG_CONNECTION = {
   user: secretsEnv.HFP_POSTGRES_USER,
   password: secretsEnv.HFP_POSTGRES_PASSWORD,
   database: secretsEnv.HFP_POSTGRES_DB,
+  ssl: true,
 }
 
 // URLs