+++ src/creators/createDeparturesResponse.ts
@@ -1,5 +1,5 @@
 import { CachedFetcher } from '../types/CachedFetcher'
-import { compact, flatten, get, groupBy, orderBy } from 'lodash'
+import { compact, flatten, get, groupBy, orderBy, uniqBy } from 'lodash'
 import { filterByDateChains } from '../utils/filterByDateChains'
 import { JoreDepartureWithOrigin, JoreStopSegment, Mode } from '../types/Jore'
 import {
@@ -28,8 +28,8 @@ import {
 } from '../utils/setCancellationsAndAlerts'
 import { Vehicles } from '../types/EventsDb'
 import { createOriginDeparture } from '../utils/createOriginDeparture'
-import { getUserGroups } from '../auth/requireUser'
 import { removeUnauthorizedData } from '../auth/removeUnauthorizedData'
+import { extraDepartureType } from '../utils/extraDepartureType'
 
 /*
   Common functions for route departures and stop departures.
@@ -230,7 +230,9 @@ export async function createDeparturesResponse(
     const groupedDepartures = groupBy<JoreDepartureWithOrigin>(
       departures,
       ({ departure_id, stop_id, day_type, extra_departure, route_id, direction }) =>
-        `${route_id}_${direction}_${departure_id}_${stop_id}_${day_type}_${extra_departure}`
+        `${route_id}_${direction}_${departure_id}_${stop_id}_${day_type}_${extraDepartureType(
+          extra_departure
+        )}`
     ) as Dictionary<JoreDepartureWithOrigin[]>
 
     const validDepartures = filterByDateChains<JoreDepartureWithOrigin>(