+++ src/creators/createJourneyResponse.ts
@@ -8,7 +8,6 @@ import { cacheFetch } from '../cache'
 import {
   AlertDistribution,
   Departure,
-  Scalars,
   ExceptionDay,
   Journey,
   JourneyCancellationEvent,
@@ -16,11 +15,22 @@ import {
   JourneyStopEvent,
   PlannedStopEvent,
   Route,
+  Scalars,
   Stop,
 } from '../types/generated/schema-types'
 import { createJourneyId } from '../utils/createJourneyId'
 import { filterByDateChains } from '../utils/filterByDateChains'
-import { compact, get, groupBy, last, orderBy, reverse } from 'lodash'
+import {
+  compact,
+  findLastIndex,
+  flatten,
+  get,
+  groupBy,
+  last,
+  mapValues,
+  orderBy,
+  reverse,
+} from 'lodash'
 import { createJourneyObject } from '../objects/createJourneyObject'
 import { getDateFromDateTime, getDepartureTime } from '../utils/time'
 import { CachedFetcher } from '../types/CachedFetcher'
@@ -35,7 +45,7 @@ import {
   setAlertsOnDeparture,
   setCancellationsOnDeparture,
 } from '../utils/setCancellationsAndAlerts'
-import { Vehicles } from '../types/EventsDb'
+import { JourneyEvents, Vehicles } from '../types/EventsDb'
 import {
   createJourneyCancellationEventObject,
   createJourneyEventObject,
@@ -52,16 +62,24 @@ import { requireVehicleAuthorization } from '../auth/requireUser'
 import { intval } from '../utils/isWithinRange'
 import { toLatLng } from '../geometry/LatLng'
 import { removeUnauthorizedData } from '../auth/removeUnauthorizedData'
+import { extraDepartureType } from '../utils/extraDepartureType'
 
 type JourneyRoute = {
   route: Route | null
   departures: Departure[]
 }
 
+type EventsType = JourneyStopEvent | JourneyEvent | PlannedStopEvent | JourneyCancellationEvent
+
 export type PlannedJourneyData = {
   departures: JoreRouteDepartureData[]
   stops: JoreStopSegment[]
 }
+
+const isPlannedEvent = (event: any): event is PlannedStopEvent => event.type === 'PLANNED'
+
+const isStopEvent = (event: any): event is PlannedStopEvent | JourneyStopEvent =>
+  typeof event.index !== 'undefined'
 
 /**
  * Fetch the journey events and filter out the invalid ones.
@@ -69,33 +87,39 @@ export type PlannedJourneyData = {
  * @param uniqueVehicleId string that identifies the requested vehicle
  * @returns Promise<Vehicles[]> the filtered events
  */
-const fetchValidJourneyEvents: CachedFetcher<Vehicles[]> = async (
+const fetchValidJourneyEvents: CachedFetcher<JourneyEvents> = async (
   fetcher,
   uniqueVehicleId
-) => {
-  const events: Vehicles[] = await fetcher()
+): Promise<JourneyEvents | false> => {
+  type JourneyEventGroup = Array<[string, Vehicles[]]>
+  type GroupedJourneyEvents = { [key: string]: JourneyEventGroup }
 
-  if (events.length === 0) {
+  const events: JourneyEvents = await fetcher()
+
+  if (flatten(Object.values(events)).length === 0) {
     return false
   }
 
   // There could have been many vehicles operating this journey. Separate them by
   // vehicle ID and use the instance argument to select the set of events.
-  const vehicleGroups = groupEventsByInstances(events || [])
-  let journeyEvents
+  const vehicleGroupedEvents: GroupedJourneyEvents = mapValues(events, (val) =>
+    groupEventsByInstances(val)
+  )
 
-  if (uniqueVehicleId) {
-    const selectedVehicleGroups = vehicleGroups.find(
-      ([groupId]) => groupId === createValidVehicleId(uniqueVehicleId)
+  let selectedVehicleId = uniqueVehicleId
+
+  // @ts-ignore typing lodash functions is impossible
+  return mapValues<JourneyEvents>(vehicleGroupedEvents, (groups: JourneyEventGroup) => {
+    if (!selectedVehicleId) {
+      selectedVehicleId = get(groups, '[0][0]', '')
+    }
+
+    const selectedGroup = groups.find(
+      ([groupId]) => groupId === createValidVehicleId(selectedVehicleId)
     )
-
-    journeyEvents = selectedVehicleGroups ? selectedVehicleGroups[1] : []
-  } else {
-    const instanceGroup = vehicleGroups[0]
-    journeyEvents = instanceGroup[1]
-  }
-
-  return journeyEvents.length !== 0 ? orderBy(journeyEvents, 'tsi', 'asc') : false
+    const selectedGroupEvents = selectedGroup ? selectedGroup[1] : []
+    return orderBy(selectedGroupEvents, 'tsi', 'asc')
+  })
 }
 
 /**
@@ -130,7 +154,7 @@ const fetchJourneyDepartures: CachedFetcher<JourneyRoute> = async (
     groupBy(
       departures,
       ({ departure_id, stop_id, day_type, extra_departure }) =>
-        `${departure_id}_${stop_id}_${day_type}_${extra_departure}`
+        `${departure_id}_${stop_id}_${day_type}_${extraDepartureType(extra_departure)}`
     ),
     date
   )
@@ -219,7 +243,7 @@ const fetchJourneyDepartures: CachedFetcher<JourneyRoute> = async (
 export async function createJourneyResponse(
   user: AuthenticatedUser | null,
   fetchRouteData: () => Promise<PlannedJourneyData>,
-  fetchJourneyEvents: () => Promise<Vehicles[]>,
+  fetchJourneyEvents: () => Promise<JourneyEvents>,
   fetchJourneyEquipment: (
     vehicleId: string | number,
     operatorId: string | number
@@ -241,7 +265,7 @@ export async function createJourneyResponse(
   // journey based on the data as the vehicle ID is part of the journey key. If an
   // ID was provided, we can make it part of the key from the start. The journey
   // key is used for caching journey events and departures.
-  function getJourneyEventsKey(events: Vehicles[] | null = []) {
+  function getJourneyEventsKey(events: JourneyEvents | null) {
     let journeyKey
 
     const journeyKeyParts = {
@@ -252,15 +276,17 @@ export async function createJourneyResponse(
       uniqueVehicleId: '',
     }
 
+    const vpEvents = events ? events.vehiclePositions : []
+
     if (uniqueVehicleId) {
       journeyKey = createJourneyId({
         ...journeyKeyParts,
         uniqueVehicleId,
       })
-    } else if (events && events.length !== 0) {
+    } else if (vpEvents.length !== 0) {
       journeyKey = createJourneyId({
         ...journeyKeyParts,
-        uniqueVehicleId: get(events, '[0].unique_vehicle_id'),
+        uniqueVehicleId: get(vpEvents, '[0].unique_vehicle_id'),
       })
     } else {
       journeyKey = createJourneyId(journeyKeyParts)
@@ -270,14 +296,14 @@ export async function createJourneyResponse(
   }
 
   // Fetch events for the journey with the cache key.
-  const journeyEvents = await cacheFetch(
-    (fetchedEvents) => {
+  const journeyEvents = await cacheFetch<JourneyEvents>(
+    (fetchedEvents: JourneyEvents) => {
       const key = getJourneyEventsKey(fetchedEvents)
       return `journey_events_${key}`
     },
     () => fetchValidJourneyEvents(fetchJourneyEvents),
     (data) => {
-      if (journeyInProgress(data)) {
+      if (journeyInProgress(data.vehiclePositions)) {
         return 1
       }
       return 24 * 60 * 60
@@ -298,14 +324,17 @@ export async function createJourneyResponse(
   )
 
   // If both of our fetches failed we'll just bail here with null.
-  if (!journeyEvents && (!routeAndDepartures || routeAndDepartures.departures.length === 0)) {
+  if (
+    journeyEvents?.vehiclePositions?.length === 0 &&
+    (!routeAndDepartures || routeAndDepartures?.departures?.length === 0)
+  ) {
     return null
   }
 
   let unsignedEvents: Vehicles[] = []
 
   const vehicleId = createValidVehicleId(
-    uniqueVehicleId || get(journeyEvents, '[0].unique_vehicle_id', '')
+    uniqueVehicleId || get(journeyEvents, 'vehiclePositions.[0].unique_vehicle_id', '')
   )
 
   let userAuthorizedForVehicle: boolean = false
@@ -386,12 +415,6 @@ export async function createJourneyResponse(
     createPlannedStopEventObject(departure, journeyAlerts)
   )
 
-  type EventsType =
-    | JourneyStopEvent
-    | JourneyEvent
-    | PlannedStopEvent
-    | JourneyCancellationEvent
-
   const stopAndCancellationEvents = orderBy<EventsType>(
     compact([...cancellationEvents, ...plannedStopEvents]),
     (event) => moment.tz(get(event, 'recordedAt', get(event, 'plannedDateTime')), TZ).unix(),
@@ -414,26 +437,12 @@ export async function createJourneyResponse(
   }
 
   // Separate the HFP events into vehicle positions, stop events and the rest of the events.
-  const { vehiclePositions = [], stopEvents = [], events = [] } = journeyEvents.reduce(
-    (groups: { [key: string]: Vehicles[] }, event) => {
-      if (event.event_type === 'VP') {
-        groups.vehiclePositions.push(event)
-      } else if (['ARS', 'DOO', 'DOC', 'DEP', 'PDE', 'PAS'].includes(event.event_type)) {
-        groups.stopEvents.push(event)
-      } else {
-        groups.events.push(event)
-      }
-
-      return groups
-    },
-    { vehiclePositions: [], stopEvents: [], events: [] }
-  )
+  const { vehiclePositions = [], stopEvents = [], otherEvents: events = [] } = journeyEvents
 
   let journeyEquipment = null
-
   const ascVehiclePositions = orderBy(vehiclePositions, 'tsi', 'asc')
 
-  if (requireVehicleAuthorization(user, vehicleId)) {
+  if (requireVehicleAuthorization(user, vehicleId) && ascVehiclePositions.length !== 0) {
     // Get the ID of the vehicle that actually operated this journey and fetch its data.
     const { owner_operator_id, vehicle_number } = ascVehiclePositions[0]
     const equipmentKey = `equipment_${owner_operator_id}_${vehicle_number}`
@@ -452,6 +461,7 @@ export async function createJourneyResponse(
   // events that we parsed from the ascVehiclePositions.
   let patchedStopEvents = [...stopEvents]
 
+  // Patch stop events by using virtual events for all stops which have no real stop events.
   for (const virtualStopEvent of virtualStopEvents) {
     const { event_type, stop } = virtualStopEvent
     const canUsePas = ['ARS', 'DEP', 'PDE'].includes(event_type)
@@ -478,8 +488,16 @@ export async function createJourneyResponse(
     evt.stop ? evt.stop + '' : evt.next_stop_id ? evt.next_stop_id + '' : 'unknown'
   )
 
-  const stopEventObjects: Array<JourneyEvent | JourneyStopEvent> = []
+  // Create event objects from other events and include cancellation events.
+  const journeyEventObjects: Array<JourneyEvent | JourneyCancellationEvent> = [
+    ...events.map((event) => createJourneyEventObject(event)),
+    ...cancellationEvents,
+  ]
 
+  const stopEventObjects: JourneyStopEvent[] = []
+
+  // Match events to departures using a variety of methods depending on what data we have available.
+  // Then create the appropriate type of JourneyEvent based on timing stop status etc.
   for (const [stopId, eventsForStop] of Object.entries(stopGroupedEvents)) {
     // Use matchedStopId to search for a stop if none is attached to the departure.
     let matchedStopId: string = ''
@@ -548,30 +566,25 @@ export async function createJourneyResponse(
       const shouldCreateStopEventObject =
         !!stop && ['ARS', useDEP ? 'DEP' : 'PDE', 'PAS'].includes(eventItem.event_type)
 
-      let eventObject: JourneyStopEvent | JourneyEvent | null = null
-
       if (!shouldCreateStopEventObject) {
-        eventObject = createJourneyEventObject(eventItem)
+        // If this should not be a stop event, put it in the journeyEventObjects array.
+        const eventObject = createJourneyEventObject(eventItem)
+        journeyEventObjects.push(eventObject)
       } else if (shouldCreateStopEventObject) {
-        eventObject = createJourneyStopEventObject(
+        // Stop events go in the stopEventObjects array.
+        const eventObject = createJourneyStopEventObject(
           eventItem,
           departure || null,
           stop,
           doorsOpened
         )
-      }
-
-      if (eventObject) {
         stopEventObjects.push(eventObject)
       }
     }
   }
 
-  // Create event objects from other events
-  const otherEvents = events.map((event) => createJourneyEventObject(event))
-
   // Exclude planned stops that did end up having events attached to them.
-  const stopsWithoutEvents = plannedStopEvents.reduce(
+  const stopsWithoutEvents: PlannedStopEvent[] = plannedStopEvents.reduce(
     (deduplicated: PlannedStopEvent[], plannedEvent) => {
       const stopWithEvent = stopEventObjects.some(
         (evt) =>
@@ -588,29 +601,57 @@ export async function createJourneyResponse(
     []
   )
 
-  const allStopEvents = [...stopEventObjects, ...stopsWithoutEvents]
+  // Planned stops should be ordered by stop order.
+  const sortedPlannedEvents: PlannedStopEvent[] = orderBy(stopsWithoutEvents, 'index')
 
-  const combinedJourneyEvents: EventsType[] = compact([
-    ...allStopEvents,
-    ...otherEvents,
-    ...cancellationEvents,
-  ])
-
-  const hasPlannedUnix = (event: any): event is PlannedStopEvent | JourneyStopEvent =>
-    typeof event.plannedUnix !== 'undefined'
-
-  const hasRecordedUnix = (event: any): event is JourneyStopEvent | JourneyEvent =>
-    typeof event.recordedAtUnix !== 'undefined'
-
-  // Combine all created event objects and order by time.
-  const sortedJourneyEvents = orderBy<EventsType>(
-    combinedJourneyEvents,
-    [
-      (event) => (hasRecordedUnix(event) ? event.recordedAtUnix : 0),
-      (event) => (hasPlannedUnix(event) ? event.plannedUnix : event.recordedAtUnix),
-    ],
-    ['asc', 'asc']
+  // Objects with observed data (ie real events) should be ordered by timestamp.
+  const sortedJourneyEvents: EventsType[] = orderBy(
+    [...stopEventObjects, ...journeyEventObjects],
+    '_sort'
   )
+
+  // Get the time info from an event.
+  const getTimeFromEvent = (event: EventsType): number => {
+    if (isPlannedEvent(event)) {
+      return event?.plannedUnix || 0
+    }
+
+    return event?.recordedAtUnix || 0
+  }
+
+  for (const plannedEvent of sortedPlannedEvents) {
+    let insertIndex = 0
+    const eventTime = plannedEvent?.plannedUnix || 0
+
+    if (plannedEvent.index === 1) {
+      // For the first stop, find the earliest index where we can insert the planned stop.
+      insertIndex = sortedJourneyEvents.findIndex(
+        (event) => getTimeFromEvent(event) >= eventTime
+      )
+
+      if (insertIndex < 1) {
+        insertIndex = 0
+      }
+    } else {
+      let prevStopIndex = findLastIndex(
+        sortedJourneyEvents,
+        (event) => isStopEvent(event) && event.index < plannedEvent.index
+      )
+
+      prevStopIndex = prevStopIndex + 1
+
+      const eventIndex = findLastIndex(
+        sortedJourneyEvents,
+        (event) => getTimeFromEvent(event) <= eventTime
+      )
+
+      insertIndex = Math.max(prevStopIndex, eventIndex)
+    }
+
+    if (insertIndex !== -1) {
+      sortedJourneyEvents.splice(insertIndex, 0, plannedEvent)
+    }
+  }
 
   let finalPositions = ascVehiclePositions
 
@@ -620,9 +661,9 @@ export async function createJourneyResponse(
     const lastEvent = last(ascVehiclePositions)
 
     // Journey start
-    const minDate = firstEvent ? intval(firstEvent.tsi) : departureDateTime.unix()
+    const minDate = firstEvent ? intval(firstEvent?.tsi || 0) : departureDateTime.unix()
     // Journey end timestamp (or scheduled start)
-    const maxDate = lastEvent ? intval(lastEvent.tsi) : departureDateTime.unix()
+    const maxDate = lastEvent ? intval(lastEvent?.tsi || 0) : departureDateTime.unix()
 
     const unsignedAroundJourney = groupBy(unsignedEvents, ({ tsi }) => {
       const intTsi = intval(tsi)