+++ src/resolvers/queryResolvers.ts
@@ -20,6 +20,7 @@ import { getCancellations } from '../getCancellations'
 import { getSettings } from '../datasources/transitlogServer'
 import { createUnsignedVehicleEventsResponse } from '../creators/createUnsignedVehicleEventsResponse'
 import { createOperatorResponse } from '../creators/createOperatorResponse'
+import { createDriverEventsResponse } from '../creators/createDriverEventsResponse'
 
 const equipment = (root, { filter, date }, { dataSources, user, skipCache }) => {
   const getEquipment = () => dataSources.JoreAPI.getEquipment()
@@ -336,6 +337,11 @@ const vehicleJourneys = (
   )
 }
 
+const driverEvents = (root, { uniqueVehicleId, date }, { dataSources, user, skipCache }) => {
+  const getDriverEvents = () => dataSources.HFPAPI.getDriverEvents(uniqueVehicleId, date)
+  return createDriverEventsResponse(user, getDriverEvents, uniqueVehicleId, date, skipCache)
+}
+
 const unsignedVehicleEvents = (
   root,
   { uniqueVehicleId, date },
@@ -425,6 +431,7 @@ export const queryResolvers: QueryResolvers = {
   weeklyDepartures,
   journey,
   vehicleJourneys,
+  driverEvents,
   unsignedVehicleEvents,
   journeys,
   journeysByBbox,