+++ src/schema.ts
@@ -6,5 +6,16 @@ import { Journey } from './schema/Journey'
 import { Equipment } from './schema/Equipment'
 import { ExceptionDays } from './schema/ExceptionDay'
 import { Alerts } from './schema/Alerts'
+import { DriverEvent } from './schema/DriverEvent'
 
-export default [Route, Stop, Departure, Journey, Equipment, ExceptionDays, Alerts, Schema]
+export default [
+  Route,
+  Stop,
+  Departure,
+  Journey,
+  DriverEvent,
+  Equipment,
+  ExceptionDays,
+  Alerts,
+  Schema,
+]