+++ src/schema/Journey.ts
@@ -15,7 +15,9 @@ export const Journey = gql`
     stopId: String
     lat: Float
     lng: Float
+    loc: String
     _isVirtual: Boolean
+    _sort: Int
   }
 
   type JourneyCancellationEvent {
@@ -33,6 +35,7 @@ export const Journey = gql`
     isCancelled: Boolean!
     cancellationType: CancellationType!
     cancellationEffect: CancellationEffect!
+    _sort: Int
   }
 
   type JourneyStopEvent {
@@ -55,14 +58,15 @@ export const Journey = gql`
     departureId: Int
     isTimingStop: Boolean!
     isOrigin: Boolean
-    index: Int
+    index: Int!
     stop: Stop
     lat: Float
     lng: Float
+    loc: String
     unplannedStop: Boolean!
     _isVirtual: Boolean
+    _sort: Int
   }
-
   type PlannedStopEvent {
     id: ID!
     type: String!
@@ -75,8 +79,9 @@ export const Journey = gql`
     departureId: Int
     isTimingStop: Boolean!
     isOrigin: Boolean
-    index: Int
+    index: Int!
     stop: Stop
+    _sort: Int
   }
 
   union JourneyEventType =
@@ -99,11 +104,13 @@ export const Journey = gql`
     vehicleId: String
     lat: Float
     lng: Float
+    loc: String
     velocity: Float
     doorStatus: Boolean
     delay: Int
     heading: Int
     mode: String
+    _sort: Int
   }
 
   type Journey {
@@ -148,6 +155,7 @@ export const Journey = gql`
     recordedAtUnix: Int!
     recordedTime: Time!
     timeDifference: Int!
+    loc: String
     alerts: [Alert!]!
     cancellations: [Cancellation!]!
     isCancelled: Boolean!