+++ src/utils/createUniqueVehicleId.ts
@@ -1,6 +1,4 @@
-import { Scalars } from '../types/generated/schema-types'
-
-export function createValidVehicleId(uniqueVehicleId: Scalars['VehicleId'] = ''): string {
+export function createValidVehicleId(uniqueVehicleId: string = ''): string {
   if (!uniqueVehicleId) {
     return ''
   }