+++ src/utils/getStopDepartureData.ts
@@ -27,7 +27,7 @@ export const getStopDepartureData = (
     return null
   }
 
-  const { tst } = departureEvent
+  const { tst, loc = null } = departureEvent
   const eventTime = parse(tst)
 
   const departureDate = !date
@@ -44,26 +44,29 @@ export const getStopDepartureData = (
     departureTime: getJourneyEventTime(departureEvent),
     departureDateTime: moment.tz(eventTime, TZ).toISOString(true),
     departureTimeDifference: departureDiff,
+    loc,
   }
 }
 
-export const getStopDepartureEvent = (events, preferDep: boolean = false) => {
+export const getStopDepartureEvent = (events, requireDep: boolean = false) => {
   if (!events || events.length === 0) {
     return null
   }
 
-  const lookForEvents = [preferDep ? 'DEP' : 'PDE', preferDep ? 'PDE' : 'DEP', 'PAS']
+  const validEventTypes = [requireDep ? 'DEP' : 'PDE', 'PAS']
   let departureEvent: Vehicles | null = null
 
-  for (const lookForType of lookForEvents) {
-    departureEvent = events.find((event) => event.event_type === lookForType)
+  for (const validType of validEventTypes) {
+    departureEvent = events.find((event) => event.event_type === validType)
 
     if (departureEvent) {
       break
+    } else {
+      departureEvent = null
     }
   }
 
-  if (!departureEvent) {
+  if (!departureEvent && !requireDep) {
     departureEvent = events[0] || null
   }